Split power supply runtime safety tests
