+++ scripts/test_assertion_allowlist.txt
@@ -11,3 +11,5 @@ src/movement_optimizer/tests/__init__.py
 src/movement_optimizer/tests/conftest.py
 src/movement_optimizer/tests/**/__init__.py
 src/movement_optimizer/tests/**/conftest.py
+# P1AM power-supply test construction helpers shared by split runtime tests.
+src/p1am_control_system/backend/tests/_power_supply_helpers.py
